Implements a frame work to test and submit a large number of different configurations. To use use the command "python python/submitBatch.py -d <directory> -m <fragment>"
This will create a directory structure rooted at <directory>
The fragment option points to some python code that will make the appropriate variations
Examples are batches/generic.py and batches/test.py

This framework functions by initializing and modifying a taggerOptions object

<fragment> must include a method named "makeJobs", it is passed a reference to the submitJob method, a reference to the saveJSON method, a refence to the taggerOptions object, and the workdirectory

Fragement will then follow its own logic to make changes to the taggerOptions object, then will pass the object and a job name to the submitJob method.
It will do this for each job to be submitted.